исправление ошибок

+++ src/lesson2/homework/HW4/Animal.java
@@ -3,9 +3,9 @@ package lesson2.homework.HW4;
 public class Animal {
     private int limitRun;
     private int limitSwim;
-    private int limitJump;
+    private float limitJump;
 
-    public Animal(int limitRun, int limitSwim, int limitJump) {
+    public Animal(int limitRun, int limitSwim, float limitJump) {
         this.limitRun = limitRun;
         this.limitSwim = limitSwim;
         this.limitJump = limitJump;
@@ -21,7 +21,7 @@ public class Animal {
         }
         return limitSwim >= distance;
     }
-    public boolean jump(int distance) {
+    public boolean jump(float distance) {
         return limitJump >= distance;
     }
 }
